// ebookstore/src/app/models/book.model.ts
export class Book {
    author_id: number;
    language: string;
    supplier: string;
    price: number;
    isbn: number;
    publish_year: string;
    subject: string;
    format: string;
    keywords: string;
    available_copies: number;
    publisher: string;
    title: string;
    image: string;
    quantity: number;
}